feat(TableHeaderLabel): Allow wrapping table header labels
This uses the non-standard
[line-clamp](https://developer.mozilla.org/en-US/docs/Web/CSS/-webkit-line-clamp)
property to restrict wrapping to just two lines, and then overflow into an ellipsis. Although
non-standard, it is supported in all major browsers except Internet
Explorer.

Unfortunately line-clamp is not possible in a button, so I've had to
change it to a div with a role of button.